RPN solve function
