+++ src/lib/scripts/calculator.ts
@@ -1,0 +1,42 @@
+import { Token } from "./dataStructures";
+
+function solveRPN(tokens: Token[]): number | string {
+  let solvedStack:Array<number | string> = [];
+  for (const token of tokens) {
+    if (token.type === "number") {
+      solvedStack.push(parseFloat(token.value));
+    }
+    else {
+      if (token.value === "-") {
+        const b = solvedStack.pop()!;
+        const a = solvedStack.pop()!;
+        solvedStack.push(Number(a) - Number(b));
+      }
+      if (token.value === "+") {
+        const b = solvedStack.pop()!;
+        const a = solvedStack.pop()!;
+        solvedStack.push(Number(a) + Number(b));
+      }
+      if (token.value === "*") {
+        const b = solvedStack.pop()!;
+        const a = solvedStack.pop()!;
+        solvedStack.push(Number(a) * Number(b));
+      }
+      if (token.value === "/") {
+        const b = solvedStack.pop()!;
+        const a = solvedStack.pop()!;
+        solvedStack.push(Number(a) / Number(b));
+      }
+      if (token.value === "^") { 
+        const b = solvedStack.pop()!;
+        const a = solvedStack.pop()!;
+        solvedStack.push(Math.pow(Number(a), Number(b)));
+      }
+      if (token.value === "_") {
+        const a = solvedStack.pop()!;
+        solvedStack.push(-Number(a));
+      }
+    }
+  }
+  return solvedStack.pop()!;
+}
