+++ src/lib/scripts/shuntingYard.ts
@@ -1,6 +1,5 @@
 import { Stack, Token, Tokens } from "./dataStructures";
 
-
 export function ShuntingYard(tokens: Tokens): Token[] | undefined {
   const output: Token[] = [];
   const stack = new Stack<Token>();
@@ -27,7 +26,7 @@ export function ShuntingYard(tokens: Tokens): Token[] | undefined {
         (
           (stack.peek()!.precedence! > token.precedence!) ||
           (stack.peek()!.precedence! === token.precedence! &&
-            token.associativity === "left") // Your existing left-associative check is fine here.
+            token.associativity === "left")
         )
       ) {
         output.push(stack.pop()!);
@@ -53,54 +52,3 @@ export function ShuntingYard(tokens: Tokens): Token[] | undefined {
   }
   return output;
 }
-
-function solve(tokens: Token[]): number | string {
-  let solvedStack:Array<number | string> = [];
-  for (const token of tokens) {
-    if (token.type === "number") {
-      solvedStack.push(parseFloat(token.value));
-    }
-    else {
-      if (token.value === "-") {
-        const b = solvedStack.pop()!;
-        const a = solvedStack.pop()!;
-        solvedStack.push(Number(a) - Number(b));
-      }
-      if (token.value === "+") {
-        const b = solvedStack.pop()!;
-        const a = solvedStack.pop()!;
-        solvedStack.push(Number(a) + Number(b));
-      }
-      if (token.value === "*") {
-        const b = solvedStack.pop()!;
-        const a = solvedStack.pop()!;
-        solvedStack.push(Number(a) * Number(b));
-      }
-      if (token.value === "/") {
-        const b = solvedStack.pop()!;
-        const a = solvedStack.pop()!;
-        solvedStack.push(Number(a) / Number(b));
-      }
-      if (token.value === "^") { 
-        const b = solvedStack.pop()!;
-        const a = solvedStack.pop()!;
-        solvedStack.push(Math.pow(Number(a), Number(b)));
-      }
-      if (token.value === "_") {
-        const a = solvedStack.pop()!;
-        solvedStack.push(-Number(a));
-      }
-    }
-  }
-  return solvedStack.pop()!;
-}
-let expression = ShuntingYard(new Tokens("-3+5*-(2-8)^2/-4"))!;
-let str = "";
-for (const token of expression) {
-  str += token.value + " ";
-}
-console.log(str);
-console.log(solve(expression));
-
-
-//  -3 5 * + 2 8 - 2 ^ -4 / -
